feat: http type helpers
