+++ playground/src/api/hello.js
@@ -1,3 +1,7 @@
+/**
+ * @param {import("adex/http").IncomingMessage} req
+ * @param {import("adex/http").ServerResponse} res
+ */
 export default (req, res) => {
   return res.json({
     pong: true,
