+++ playground/src/api/html.js
@@ -1,3 +1,7 @@
+/**
+ * @param {import("adex/http").IncomingMessage} req
+ * @param {import("adex/http").ServerResponse} res
+ */
 export default (req, res) => {
   return res.html(`<h1>Html Response</h1>`)
 }
